Refactor threads cpputest to gtest and add a tss test
