+++ libs/utils/CMakeLists.txt
@@ -128,10 +128,6 @@ if (ENABLE_TESTING)
         target_include_directories(array_list_test PRIVATE include_deprecated)
         target_link_libraries(array_list_test  Celix::utils CppUTest::CppUTest pthread)
 
-        add_executable(celix_threads_test private/test/celix_threads_test.cpp)
-        target_include_directories(celix_threads_test PRIVATE include_deprecated)
-        target_link_libraries(celix_threads_test Celix::utils CppUTest::CppUTest ${CppUTest_EXT_LIBRARIES} pthread)
-
         add_executable(linked_list_test private/test/linked_list_test.cpp)
         target_include_directories(linked_list_test PRIVATE include_deprecated)
         target_link_libraries(linked_list_test  Celix::utils CppUTest::CppUTest pthread)
@@ -160,7 +156,6 @@ if (ENABLE_TESTING)
 
         add_test(NAME run_array_list_test COMMAND array_list_test)
         add_test(NAME run_hash_map_test COMMAND hash_map_test)
-        add_test(NAME run_celix_threads_test COMMAND celix_threads_test)
         add_test(NAME run_linked_list_test COMMAND linked_list_test)
         add_test(NAME run_properties_test COMMAND properties_test)
         add_test(NAME run_ip_utils_test COMMAND ip_utils_test)
@@ -168,7 +163,6 @@ if (ENABLE_TESTING)
 
         setup_target_for_coverage(array_list_test)
         setup_target_for_coverage(hash_map_test)
-        setup_target_for_coverage(celix_threads_test)
         setup_target_for_coverage(linked_list_test)
         setup_target_for_coverage(properties_test)
         setup_target_for_coverage(ip_utils_test)
